Fix errors without ecto
When running on a project that does not include Ecto a compilation error would
be returned:

```
** (MatchError) no match of right hand side value: {:error, :nofile}
    lib/my_app/enum_container.ex:10: (module)
    lib/my_app/enum_container.ex:8: (module)
    (stdlib 3.12.1) erl_eval.erl:680: :erl_eval.do_apply/6
    (elixir 1.10.2) lib/kernel/parallel_compiler.ex:304: anonymous fn/4 in Kernel.ParallelCompiler.spawn_workers/7
```

This is because `if {:module, _module} = Code.ensure_compiled(Ecto.Type) do`
results in a `MatchError` because the left side does not match the right. The
fix is to switch to a `with` statement that does not execute the body in the
face of a match error.

I also added a `@moduledoc` for `EnumType` because `EnumType` is not a private
module since it is meant to be called from our own modules.

You're likely already familiary with it, but here is where the meaning of
`@moduledoc false`:
https://hexdocs.pm/elixir/writing-documentation.html#hiding-internal-modules-and-functions

I also added `generated: true` and `location: :keep` to most of the quote
blocks. They mark the code as generated which could ignore any generated
warnings and provide more helpful stack trace info:
https://hexdocs.pm/elixir/master/Kernel.SpecialForms.html#quote/2-options